# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Adapta API
  description: |
    **Your model. Your data. Your servers.**

    Adapta is a self-hosted model-customization platform. Technical teams deploy
    it on their own infrastructure to specialize and serve private language models two ways,
    each behind an OpenAI-compatible API. Your data never leaves your servers.

    **Give it knowledge (RAG)** — upload documents, embed them, and get answers grounded in
    your content, with citations. Weights never change. CPU-only.

    **Change how it behaves (fine-tuning / LoRA)** — train a QLoRA adapter on instruction
    data; the adapter only serves once it passes the eval gate (score ≥ 0.6 or a clear
    improvement over the base model). Requires a GPU.

    **Composed endpoints** — a single project endpoint can serve both: retrieved context
    with citations (RAG) and trained tone/format (the LoRA adapter) in one call.

    ---

    **Auth:** management endpoints require a JWT Bearer token from `POST /v1/auth/login`.
    The `POST /v1/chat/completions` endpoint uses a scoped project API key (`adp_*` prefix).

  version: 0.1.0
  license:
    name: MIT
    url: https://github.com/yielab/adapta/blob/main/LICENSE
  contact:
    name: Adapta
    url: https://github.com/yielab/adapta
  x-logo:
    url: /console/logo.svg
    altText: Adapta
servers:
  - url: /v1
    description: Versioned API
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects for a team
      security: [{BearerAuth: []}]
      parameters:
        - name: team_id
          in: query
          required: true
          schema: {type: string}
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '200': {description: Projects, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/ProjectResponse'}}}}}

    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      security: [{BearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ProjectCreate'}
      responses:
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/ProjectResponse'}}}}
  /projects/{project_id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project
      security: [{BearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ProjectId'}]
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '200': {description: Project, content: {application/json: {schema: {$ref: '#/components/schemas/ProjectResponse'}}}}
        '404': {$ref: '#/components/responses/Error'}

    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project (admin only)
      security: [{BearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ProjectId'}]
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '204': {description: Deleted}

  # ──────────────────────────────────────────────────────────────────────────
  # Files (RAG)
  # ──────────────────────────────────────────────────────────────────────────
...
  /projects/{project_id}/usage:
    get:
      operationId: getUsage
      tags: [usage]
      summary: Daily token-usage rollup for the project's endpoint
      security: [{BearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ProjectId'}]
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '200': {description: Usage, content: {application/json: {schema: {$ref: '#/components/schemas/UsageResponse'}}}}

  # ──────────────────────────────────────────────────────────────────────────
  # Platform settings (DB-backed, whitelisted, org-scoped overrides)
  # ──────────────────────────────────────────────────────────────────────────
  /settings:
    get:
      operationId: listSettings
      tags: [settings]
      summary: List whitelisted settings with their effective value and provenance
      security: [{BearerAuth: []}]
      parameters:
        - name: team_id
          in: query
          required: true
          schema: {type: string}
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '200': {description: Settings, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/SettingResponse'}}}}}

    put:
      operationId: putSetting
      tags: [settings]
      summary: Override a whitelisted setting (admin only)
      security: [{BearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SettingWriteRequest'}
      responses:
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '200': {description: Setting written, content: {application/json: {schema: {$ref: '#/components/schemas/SettingWriteResult'}}}}
  /settings/{key}:
    delete:
      operationId: deleteSetting
      tags: [settings]
      summary: Reset a setting to its default by removing the override (admin only)
      security: [{BearerAuth: []}]
      parameters:
        - {name: key, in: path, required: true, schema: {type: string}}
        - {name: team_id, in: query, required: true, schema: {type: string}}
      responses:
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '204': {description: Reset to default}

  # ──────────────────────────────────────────────────────────────────────────
  # API keys
  # ──────────────────────────────────────────────────────────────────────────
...
  /projects/{project_id}/datasets/synthesize:
    post:
      operationId: synthesizeDataset
      tags: [datasets]
      summary: Synthesize an instruction dataset from indexed documents
      description: |
        Kicks off a background job that reads the project's ChromaDB collection,
        calls the local inference engine to generate question/answer pairs, and
        writes a validated JSONL dataset. Poll the returned dataset_id via
        GET /v1/projects/{project_id}/datasets/{dataset_id} for status.
        Only available for `finetune` projects that already have indexed files.
      security: [{BearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/ProjectId'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SynthesizeRequest'}
      responses:
        '404': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '202': {description: Synthesis started, content: {application/json: {schema: {$ref: '#/components/schemas/SynthesizeResponse'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}

  # ──────────────────────────────────────────────────────────────────────────
  # Chat (OpenAI-compatible, scoped key auth)
  # ──────────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: Scoped project API key (adp_...)
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema: {type: string}
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
  schemas:
...
    ProjectType:
      type: string
      enum: [rag, finetune]
...
    ProjectCreate:
      type: object
      required: [name, type, base_model, team_id]
      properties:
        # maxLength mirrors projects.name String(128) — see RegisterRequest note.
        name: {type: string, minLength: 1, maxLength: 128}
        type: {$ref: '#/components/schemas/ProjectType'}
        base_model: {type: string}
        description: {type: string, nullable: true}
        team_id: {type: string}
    ProjectResponse:
      type: object
      required: [id, name, type, status, base_model, description, team_id, created_at, summary]
      properties:
        id: {type: string}
        name: {type: string}
        type: {$ref: '#/components/schemas/ProjectType'}
        status: {type: string}
        base_model: {type: string}
        description: {type: string, nullable: true}
        team_id: {type: string}
        created_at: {type: string}
        # Dashboard read-model (§C2.1) — null on create, populated on list/get.
        summary: {$ref: '#/components/schemas/ProjectSummary', nullable: true}

    # Project dashboard read-model (computed aggregate served with each project).
    FileSummary:
      type: object
      required: [total, indexed, chunks]
      properties:
        total: {type: integer}
        indexed: {type: integer}
        chunks: {type: integer}
    DatasetSummary:
      type: object
      required: [total, valid]
      properties:
        total: {type: integer}
        valid: {type: integer}
    JobSummary:
      type: object
      required: [total, running]
      properties:
        total: {type: integer}
        running: {type: integer}
        last_status: {type: string, nullable: true}
        last_eval_score: {type: number, nullable: true}
        gate_passed: {type: boolean, nullable: true}
    EndpointSummary:
      type: object
      required: [exists]
      properties:
        exists: {type: boolean}
        slug: {type: string, nullable: true}
        status: {type: string, nullable: true}
    UsageSummary:
      type: object
      required: [requests, total_tokens]
      properties:
        requests: {type: integer}
        total_tokens: {type: integer}
    ProjectSummary:
      type: object
      required: [stage, files, datasets, jobs, endpoint, keys_active, usage_7d]
      properties:
        stage: {type: string}
        files: {$ref: '#/components/schemas/FileSummary'}
        datasets: {$ref: '#/components/schemas/DatasetSummary'}
        jobs: {$ref: '#/components/schemas/JobSummary'}
        endpoint: {$ref: '#/components/schemas/EndpointSummary'}
        keys_active: {type: integer}
        usage_7d: {$ref: '#/components/schemas/UsageSummary'}
        last_activity_at: {type: string, nullable: true}

    # Files
...
    UsageDay:
      type: object
      required: [day, prompt_tokens, completion_tokens, total_tokens, request_count]
      properties:
        day: {type: string}
        prompt_tokens: {type: integer}
        completion_tokens: {type: integer}
        total_tokens: {type: integer}
        request_count: {type: integer}
    UsageResponse:
      type: object
      required: [project_id, total_prompt_tokens, total_completion_tokens, total_tokens, total_requests, days]
      properties:
        project_id: {type: string}
        total_prompt_tokens: {type: integer}
        total_completion_tokens: {type: integer}
        total_tokens: {type: integer}
        total_requests: {type: integer}
        days: {type: array, items: {$ref: '#/components/schemas/UsageDay'}}

    # Keys
...
    SynthesizeRequest:
      type: object
      properties:
        n_pairs_per_chunk: {type: integer, minimum: 1, maximum: 10, default: 3}
        max_chunks: {type: integer, minimum: 1, maximum: 500, default: 50}
        system_prompt: {type: string, nullable: true, maxLength: 1000}
        base_model: {type: string, nullable: true}
    SynthesizeResponse:
      type: object
      required: [dataset_id, status, message]
      properties:
        dataset_id: {type: string}
        status: {type: string}
        message: {type: string}

    # Settings (DB-backed whitelisted org-scoped overrides). `value`/`default`/
    # `min`/`max` are left untyped because a setting is int OR float depending on
    # its key; an untyped schema preserves the original int (a `number` type
    # would coerce 512 → 512.0 on the way out).
    SettingResponse:
      type: object
      required: [key, label, group, value, source, default, min, max, description]
      properties:
        key: {type: string}
        label: {type: string}
        group: {type: string}
        value: {}
        source: {type: string}
        default: {}
        min: {}
        max: {}
        description: {type: string}
    SettingWriteRequest:
      type: object
      required: [team_id, key, value]
      properties:
        team_id: {type: string}
        key: {type: string}
        value: {}
    SettingWriteResult:
      type: object
      required: [key, value, source, default]
      properties:
        key: {type: string}
        value: {}
        source: {type: string}
        default: {}

    # Models catalog
...
    Error:
      type: object
      properties:
        code: {type: string}
        message: {type: string}
        correlation_id: {type: string, nullable: true}
    ErrorResponse:
      type: object
      properties:
        error: {$ref: '#/components/schemas/Error'}